move DatabaseMiddleware to peripherals
